Player stats have been finally removed
script still exists but everything has been moved into their respected classes and added health comp so VFX are not auto.
