--- Assets/Programming/Character/PlayerStats.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 65bcd4e1142ca2a4a8e3dce38f7b3570
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
