+++ Assets/Programming/UI/Player/PlayerReferenceFetcher.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fee6ef1fa45e31b4a83bc1491744b5b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
